Big bindings rework. From now on, bindings between shader and config are matched by order

+++ src/renderer/renderer_resource_bind_group_layout.odin
@@ -78,7 +78,7 @@ BindGroupLayoutDesc :: struct {
 BindGroupLayoutResource :: struct {
 	desc:                BindGroupLayoutDesc,
 	hash:                u32,
-	num_dynamic_offsets: u8,
+	num_dynamic_offsets: u32,
 }
 
 //---------------------------------------------------------------------------//
